Update information indexed for Transfers
refs #6131

All files in a Transfer will be indexed so that the metadata and logs will
be available to be viewed in the SIP Arrangement panel.  Indexing and
updating the index on a file do not verify that they are in the DB anymore.
The full file path is not indexed, but the relative path (including the
transfer folder) is.

Add a wrapper for search_raw that queries for MAX_QUERY_SIZE entries from
ElasticSearch.  The default number returned is 10, and there is no way to
request all entries that match a search.