Arranged files are not draggable
refs #6131

A file should not be multiple SIPs.  If a file has been arranged (can be
found in SIPArrange.original_files), then it is greyed out in the originals
panel. When a file is dragged to the arranged panel, or deleted from the
arranged panel, the originals panel refreshes to get the updated status.

SIPArrange original paths are now unique, and when copying a folder, does
not copy files from originals that have already been arranged.